requirements.txt
requests
qrcode[pil]
google-api-python-client
google-auth
google-auth-oauthlib
